Improve tree_client to receive several inputs

+++ include/client/client_stub-private.h
@@ -9,7 +9,7 @@
 
 struct rtree_t {
   char* server_ip_address;
-  short server_port;
+  int server_port;
   int sockfd;
 };
